excise pure state mutation functions into helper
