+++ gs_desktop_application/src/main/scala/dev/nateschieber/groovesprings/actors/GsAppStateManager.scala
@@ -11,6 +11,7 @@ import dev.nateschieber.groovesprings.enums.GsAppStateManagerTimer.currFrameIdCa
 import dev.nateschieber.groovesprings.enums.GsLoopType.{ALL, ONE}
 import dev.nateschieber.groovesprings.enums.{GsAppStateManagerTimer, GsLoopType, GsPlayState, GsPlaybackSpeed}
 import dev.nateschieber.groovesprings.enums.GsPlayState.{PAUSE, PLAY, STOP}
+import dev.nateschieber.groovesprings.helpers.GsAppStateMutations
 import dev.nateschieber.groovesprings.rest.FileSelectJsonSupport
 import dev.nateschieber.groovesprings.traits.{AddTrackToPlaylist, ClearPlaylist, CurrPlaylistTrackIdx, GsCommand, HydrateState, HydrateStateToDisplay, InitialTrackSelect, NextOrPrevTrack, NextTrack, PauseTrig, PlayFromNextOrPrevTrack, PlayFromTrackSelectTrig, PlayTrig, PrevTrack, ReadPlaybackThreadState, RespondAddTrackToPlaylist, RespondCurrPlaylistTrackIdx, RespondHydrateState, RespondNextOrPrevTrack, RespondPauseTrig, RespondPlayFromNextOrPrevTrack, RespondPlayFromTrackSelectTrig, RespondPlayTrig, RespondPlaybackThreadState, RespondRestTrackSelect, RespondSetPlaylist, RespondStopTrig, RespondTimerStart, RespondTrackSelect, RestTrackSelect, SendLastFrameId, SendReadComplete, SetFilePath, SetLoopType, SetPlaybackSpeed, SetPlaylist, SetShuffle, StopTrig, TimerStart, TrackSelect, TransportTrig}
 
@@ -77,6 +78,8 @@ class GsAppStateManager(
   extends AbstractBehavior[GsCommand](context)
     with FileSelectJsonSupport with AppStateJsonSupport {
 
+  private val stateMutation: GsAppStateMutations = GsAppStateMutations()
+
   private var currFrameIdCacheTimerRef: ActorRef[GsCommand] = context.spawn(
     GsTimer(GsAppStateManagerTimer.currFrameIdCache.id),
     UUID.randomUUID().toString
@@ -117,270 +120,6 @@ class GsAppStateManager(
     )
   }
 
-  def setCurrTrack(appState: AppState, track: Track): AppState = {
-    AppState(
-      appState.playState,
-      appState.playbackSpeed,
-      appState.loopType,
-      appState.shuffle,
-      appState.currFrameId,
-      track,
-      appState.currPlaylistTrackIdx,
-      appState.playlist
-    )
-  }
-
-  def addTrackToPlaylist(appState: AppState, track: Track): AppState = {
-    AppState(
-      appState.playState,
-      appState.playbackSpeed,
-      appState.loopType,
-      appState.shuffle,
-      appState.currFrameId,
-      appState.currTrack,
-      appState.currPlaylistTrackIdx,
-      Playlist(appState.playlist.id, appState.playlist.name, appState.playlist.tracks ++ List(track))
-    )
-  }
-
-  private def setPlaylist(appState: AppState, playlist: Playlist): AppState = {
-    AppState(
-      appState.playState,
-      appState.playbackSpeed,
-      appState.loopType,
-      appState.shuffle,
-      appState.currFrameId,
-      appState.currTrack,
-      0,
-      playlist
-    )
-  }
-
-  private def clearPlaylist(appState: AppState): AppState = {
-    AppState(
-      appState.playState,
-      appState.playbackSpeed,
-      appState.loopType,
-      appState.shuffle,
-      appState.currFrameId,
-      appState.currTrack,
-      appState.currPlaylistTrackIdx,
-      EmptyPlaylist
-    )
-  }
-
-  private def newRandomIdx(playlistLength: Int, oldIdx: Int): Int = {
-    var randomIdx: Int = Random.between(0, playlistLength)
-    while
-      randomIdx == oldIdx
-    do randomIdx = newRandomIdx(playlistLength, oldIdx)
-    randomIdx
-  }
-
-  // TODO:
-  //   - set new currFrameId based on playbackSpeed when moving between tracks
-  private def prevTrack(appState: AppState): AppState = {
-    if (!(appState.playlist.tracks.map(track => track.id) contains appState.currTrack.id))
-      return appState
-
-    val playlistLength = appState.playlist.tracks.length
-    if (playlistLength == 0)
-      return AppState(
-        appState.playState,
-        appState.playbackSpeed,
-        appState.loopType,
-        appState.shuffle,
-        0,
-        appState.currTrack,
-        0,
-        EmptyPlaylist
-      )
-
-    val oldIdx = appState.currPlaylistTrackIdx
-    val newIdx = if (appState.loopType == GsLoopType.ONE)
-                   oldIdx
-                 else if (appState.shuffle)
-                   newRandomIdx(appState.playlist.tracks.length, oldIdx)
-                 else
-                   (oldIdx + (playlistLength - 1)) % playlistLength
-
-
-    AppState(
-      appState.playState,
-      appState.playbackSpeed,
-      appState.loopType,
-      appState.shuffle,
-      0,
-      appState.playlist.tracks(newIdx),
-      newIdx,
-      appState.playlist
-    )
-  }
-
-  private def nextTrack(appState: AppState): AppState = {
-    if (!(appState.playlist.tracks.map(track => track.id) contains appState.currTrack.id))
-      return appState
-
-    if (appState.playlist.tracks.isEmpty)
-      return AppState(
-        appState.playState,
-        appState.playbackSpeed,
-        appState.loopType,
-        appState.shuffle,
-        0,
-        appState.currTrack,
-        0,
-        EmptyPlaylist
-      )
-
-    val oldIdx = appState.currPlaylistTrackIdx
-    // loop behavior
-    if (appState.loopType == GsLoopType.NONE
-          && oldIdx == appState.playlist.tracks.length - 1)
-      return AppState(
-        GsPlayState.STOP,
-        appState.playbackSpeed,
-        appState.loopType,
-        appState.shuffle,
-        0,
-        appState.playlist.tracks.head,
-        0,
-        appState.playlist
-      )
-
-    val newIdx = if (appState.loopType == GsLoopType.ONE)
-                   oldIdx
-                 else if (appState.shuffle)
-                   newRandomIdx(appState.playlist.tracks.length, oldIdx)
-                 else
-                   (oldIdx + 1) % appState.playlist.tracks.length
-
-    AppState(
-      appState.playState,
-      appState.playbackSpeed,
-      appState.loopType,
-      appState.shuffle,
-      0,
-      appState.playlist.tracks(newIdx),
-      newIdx,
-      appState.playlist
-    )
-  }
-
-  private def setPlayState(appState: AppState, newPlayState: GsPlayState): AppState = {
-    AppState(
-      newPlayState,
-      appState.playbackSpeed,
-      appState.loopType,
-      appState.shuffle,
-      appState.currFrameId,
-      appState.currTrack,
-      appState.currPlaylistTrackIdx,
-      appState.playlist
-    )
-  }
-
-  private def setPlaybackSpeed(state: AppState, newPlaybackSpeed: GsPlaybackSpeed): AppState = {
-    AppState(
-      state.playState,
-      newPlaybackSpeed,
-      state.loopType,
-      state.shuffle,
-      state.currFrameId,
-      state.currTrack,
-      state.currPlaylistTrackIdx,
-      state.playlist
-    )
-  }
-
-  private def setShuffle(state: AppState): AppState = {
-    AppState(
-      state.playState,
-      state.playbackSpeed,
-      state.loopType,
-      !state.shuffle,
-      state.currFrameId,
-      state.currTrack,
-      state.currPlaylistTrackIdx,
-      state.playlist
-    )
-  }
-
-  private def setLoopType(state: AppState): AppState = {
-    val newLoopType = appState.loopType match {
-      case GsLoopType.ONE  => GsLoopType.NONE
-      case GsLoopType.ALL  => GsLoopType.ONE
-      case GsLoopType.NONE => GsLoopType.ALL
-    }
-
-    AppState(
-      state.playState,
-      state.playbackSpeed,
-      newLoopType,
-      state.shuffle,
-      state.currFrameId,
-      state.currTrack,
-      state.currPlaylistTrackIdx,
-      state.playlist
-    )
-  }
-
-  private def setCurrPlaylistTrackIdx(appState: AppState, newIdx: Int): AppState = {
-    var optTrack = appState.playlist.tracks.lift(newIdx)
-    if (optTrack.isDefined)
-      AppState(
-        appState.playState,
-        appState.playbackSpeed,
-        appState.loopType,
-        appState.shuffle,
-        appState.currFrameId,
-        optTrack.get,
-        newIdx,
-        appState.playlist)
-    else
-      AppState(
-        appState.playState,
-        appState.playbackSpeed,
-        appState.loopType,
-        appState.shuffle,
-        appState.currFrameId,
-        EmptyTrack,
-        0,
-        appState.playlist
-      )
-  }
-
-  private def setCurrFrameId(newCurrFrameId: Long): Unit = {
-    setAppState(
-      AppState(
-        appState.playState,
-        appState.playbackSpeed,
-        appState.loopType,
-        appState.shuffle,
-        newCurrFrameId,
-        appState.currTrack,
-        appState.currPlaylistTrackIdx,
-        appState.playlist
-      )
-    )
-  }
-
-  private def cacheCurrFrameId(): Unit = {
-    val currFrameId = if (appState.playState == GsPlayState.STOP) 0L else GsPlaybackThread.getCurrFrameId.asInstanceOf[Long]
-    setAppState(
-      AppState(
-        appState.playState,
-        appState.playbackSpeed,
-        appState.loopType,
-        appState.shuffle,
-        currFrameId,
-        appState.currTrack,
-        appState.currPlaylistTrackIdx,
-        appState.playlist
-      )
-    )
-  }
-
   private def hydrateState(): Unit = {
     gsDisplayRef ! HydrateState(appState.toJson.compactPrint, context.self)
   }
@@ -406,7 +145,7 @@ class GsAppStateManager(
     msg match {
       case RestTrackSelect(track, replyTo) =>
         setAppState(
-          setCurrTrack(appState, track)
+          stateMutation.setCurrTrack(appState, track)
         )
         gsPlaybackRef ! TrackSelect(track, context.self)
         hydrateState()
@@ -415,7 +154,7 @@ class GsAppStateManager(
 
       case RespondTrackSelect(path, _replyTo) =>
         setAppState(
-          setPlayState(appState, GsPlayState.PLAY)
+          stateMutation.setPlayState(appState, GsPlayState.PLAY)
         )
         gsPlaybackRef ! PlayFromTrackSelectTrig(path, context.self)
         hydrateState()
@@ -442,7 +181,7 @@ class GsAppStateManager(
 
       case AddTrackToPlaylist(track, replyTo) =>
         setAppState(
-          addTrackToPlaylist(appState, track)
+          stateMutation.addTrackToPlaylist(appState, track)
         )
         hydrateState()
         replyTo ! RespondAddTrackToPlaylist(context.self)
@@ -450,7 +189,7 @@ class GsAppStateManager(
 
       case SetPlaylist(playlist, replyTo) =>
         setAppState(
-          setPlaylist(appState, playlist)
+          stateMutation.setPlaylist(appState, playlist)
         )
         hydrateState()
         replyTo ! RespondSetPlaylist(context.self)
@@ -458,14 +197,14 @@ class GsAppStateManager(
 
       case ClearPlaylist(replyTo) =>
         setAppState(
-          clearPlaylist(appState)
+          stateMutation.clearPlaylist(appState)
         )
         hydrateState()
         Behaviors.same
 
       case CurrPlaylistTrackIdx(newIdx, replyTo) =>
         setAppState(
-          setCurrPlaylistTrackIdx(appState, newIdx)
+          stateMutation.setCurrPlaylistTrackIdx(appState, newIdx)
         )
         gsPlaybackRef ! TrackSelect(appState.currTrack, context.self)
         hydrateState()
@@ -486,7 +225,7 @@ class GsAppStateManager(
 
       case RespondPlayTrig(replyTo) =>
         setAppState(
-          setPlayState(appState, GsPlayState.PLAY)
+          stateMutation.setPlayState(appState, GsPlayState.PLAY)
         )
         replyTo ! ReadPlaybackThreadState(context.self)
         hydrateState()
@@ -502,12 +241,12 @@ class GsAppStateManager(
           if (appState.playState == GsPlayState.PLAY)
             if (appState.playbackSpeed.value > 0)
               setAppState(
-                nextTrack(appState)
+                stateMutation.nextTrack(appState)
               )
             else if (appState.playbackSpeed.value < 0)
               // TODO: start from end of track
               setAppState(
-                prevTrack(appState)
+                stateMutation.prevTrack(appState)
               )
 
           postNextOrPrevTrackStateUpdate()
@@ -523,21 +262,21 @@ class GsAppStateManager(
 
       case RespondPauseTrig(replyTo) =>
         setAppState(
-          setPlayState(appState, GsPlayState.PAUSE)
+          stateMutation.setPlayState(appState, GsPlayState.PAUSE)
         )
         hydrateState()
         Behaviors.same
 
       case RespondStopTrig(replyTo) =>
         setAppState(
-          setPlayState(appState, GsPlayState.STOP)
+          stateMutation.setPlayState(appState, GsPlayState.STOP)
         )
         hydrateState()
         Behaviors.same
 
       case SetPlaybackSpeed(newPlaybackSpeed, _) =>
         setAppState(
-          setPlaybackSpeed(appState, newPlaybackSpeed)
+          stateMutation.setPlaybackSpeed(appState, newPlaybackSpeed)
         )
         gsPlaybackRef ! SetPlaybackSpeed(newPlaybackSpeed, gsDisplayRef)
         hydrateState()
@@ -545,28 +284,28 @@ class GsAppStateManager(
 
       case SetLoopType(replyTo) =>
         setAppState(
-          setLoopType(appState)
+          stateMutation.setLoopType(appState)
         )
         hydrateState()
         Behaviors.same
 
       case SetShuffle(replyTo) =>
         setAppState(
-          setShuffle(appState)
+          stateMutation.setShuffle(appState)
         )
         hydrateState()
         Behaviors.same
 
       case PrevTrack() =>
         setAppState(
-          prevTrack(appState)
+          stateMutation.prevTrack(appState)
         )
         postNextOrPrevTrackStateUpdate()
         Behaviors.same
 
       case NextTrack() =>
         setAppState(
-          nextTrack(appState)
+          stateMutation.nextTrack(appState)
         )
         postNextOrPrevTrackStateUpdate()
         Behaviors.same
@@ -575,11 +314,16 @@ class GsAppStateManager(
         val timer: GsAppStateManagerTimer = GsAppStateManagerTimer.fromId(timerId)
         timer match {
           case GsAppStateManagerTimer.currFrameIdCache =>
-            cacheCurrFrameId()
+            val currFrameId = if (appState.playState == GsPlayState.STOP) 0L else GsPlaybackThread.getCurrFrameId.asInstanceOf[Long]
+            setAppState(
+              stateMutation.setCurrFrameId(appState, currFrameId)
+            )
             if (appState.playState == GsPlayState.PLAY)
               currFrameIdCacheTimerStart()
             else if (appState.playState == GsPlayState.STOP)
-            setCurrFrameId(0)
+            setAppState(
+              stateMutation.setCurrFrameId(appState, 0)
+            )
 
           case GsAppStateManagerTimer.playbackThreadPoll =>
             gsPlaybackRef ! ReadPlaybackThreadState(context.self)
